+++ simulator/placing_strategy/MinTaskCount.h
@@ -17,7 +17,7 @@ struct MinTaskCount : public IPlacingStrategy {
         }
 
         unsigned bestMachineIndex{machines[0]->machineIndex_};
-        size_t minTaskCount{machines[0]->currentTaskCount_};
+        unsigned minTaskCount{machines[0]->currentTaskCount_};
 
         for (size_t i = 1; i < split; ++i) {
             const Machine* machine = machines[i];
